заменил деньги на bigdecimal и добавил примитивный логер

+++ src/main/java/ru/bulkashmak/api/models/user/UserRequest.java
@@ -2,6 +2,8 @@ package ru.bulkashmak.api.models.user;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,7 +13,7 @@ public class UserRequest {
     private String secondName;
     private Integer age;
     private UserResponse.Sex sex;
-    private Double money;
+    private BigDecimal money;
 
     @Getter
     enum Sex {
